feat(ios): Remove Appdelegate
